backend/requirements.txt
flask
flask-cors
flask-session
python-dotenv
openai
requests
google-auth
google-auth-oauthlib
